+++ BlueprintCoreGen/Templates/ConditionsBuilder/MiscEx/ConditionsBuilderMiscEx.cs
@@ -1,0 +1,18 @@
+﻿namespace BlueprintCoreGen.Conditions.Builder.MiscEx
+{
+  /// <summary>
+  /// Extension to <see cref="ConditionsBuilder"/> for conditions without a better extension container such as
+  /// difficulty.
+  /// </summary>
+  /// <inheritdoc cref="ConditionsBuilder"/>
+  public static class ConditionsBuilderMiscEx
+  {
+    //----- Auto Generated -----//
+
+    // [Generate(Kingmaker.UnitLogic.Mechanics.Conditions.ContextConditionDifficultyHigherThan)]
+    // [Generate(Kingmaker.UnitLogic.Mechanics.Conditions.DifficultyHigherThan)]
+    // [Generate(Kingmaker.UnitLogic.Mechanics.Conditions.EnlargedEncountersCapacity)]
+    // [Generate(Kingmaker.Assets.Designers.EventConditionActionSystem.Conditions.Paused)]
+
+  }
+}